refactor: 배럴 익스포트 관련 수정
