+++ src/shared/api/index.ts
@@ -1,3 +1,6 @@
 export { axiosInstance } from './axiosInstance';
 export { ApiError } from './apiError';
-export * from './query';
+
+export { createQueryClient, enableReactQueryDevTools } from './query/client';
+export { ReactQueryProvider } from './query/provider';
+export { HydrationProvider } from './query/hydration';
