--- src/shared/api/query/index.ts
@@ -1,3 +1,0 @@
-export { createQueryClient, enableReactQueryDevTools } from './client';
-export { ReactQueryProvider } from './provider';
-export { HydrationProvider } from './hydration';
