+++ src/shared/config/index.ts
@@ -1,2 +1,2 @@
 // 설정 배럴 익스포트
-export * from './api';
+export { API_CONFIG, API_ENDPOINTS, HTTP_STATUS, ERROR_CODES } from './api';
