+++ src/shared/lib/index.ts
@@ -1,2 +1,2 @@
 // Shared 라이브러리 배럴 익스포트
-export * from './store';
+export { cn } from './cn';
