--- src/shared/types/index.ts
@@ -1,3 +1,0 @@
-// 타입 배럴 익스포트
-// export * from './api';
-// export * from './common';
