+++ src/shared/ui/button/Button.tsx
@@ -1,5 +1,5 @@
 import React, { ButtonHTMLAttributes } from 'react';
-import { cn } from '@/shared/lib/cn';
+import { cn } from '@/shared/lib';
 import { VariantProps, cva } from 'class-variance-authority';
 
 export interface ButtonProps
