--- src/shared/ui/index.ts
@@ -1,2 +1,0 @@
-// UI 컴포넌트 배럴 익스포트
-export * from './icon';
